move rebalance related test out of CoordinatorEventProcessorTest
AI-Model: gpt-5.6-sol
AI-Contributed/Feature: 0/0
AI-Contributed/UT: 1670/1670
